+++ backupscripts/readinglist.py
@@ -1,11 +1,16 @@
 #!/usr/bin/python3
 # -*- coding: utf-8 -*-
 
-# Copyright © 2014 Martin Ueding <dev@martin-ueding.de>
+# Copyright © 2014-2015 Martin Ueding <dev@martin-ueding.de>
 
 import configparser
-import os.path
 import glob
+import os.path
+import re
+import shutil
+
+import lxml.html
+import requests
 
 DIR = os.path.expanduser('~/Leseliste')
 
@@ -22,9 +27,24 @@ def get_urls():
             if key.startswith('url'):
                 urls.append(parser['Desktop Entry'][key])
 
+                rename_file_if_needed(file_, parser['Desktop Entry'][key])
+
     urls.sort()
 
     return urls
+
+
+def rename_file_if_needed(filename, url):
+    if not 'watch?v=' in filename:
+        return
+
+    r = requests.get(url)
+    t = lxml.html.fromstring(r.text)
+    title = t.find(".//title").text
+    title_clean = re.sub(r'[^A-Za-z0-9-_]+', '_', title)
+
+    shutil.move(filename, os.path.join(os.path.dirname(filename), title_clean[:100] + '.desktop'))
+
 
 def main():
     with open(os.path.join(DIR, 'websites.html'), 'w') as handle: